@@ -1,6 +1,6 @@
 module knot8.io
 
-go 1.19
+go 1.21
 
 require (
 	github.com/alecthomas/kong v0.7.1
